// web/src/gen/types/Badge.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type Badge = {
  /**
   * @type string | undefined
   */
  color?: string
  /**
   * @type string | undefined
   */
  created_at?: string
  /**
   * @type string | undefined
   */
  icon?: string
  /**
   * @type integer | undefined
   */
  id?: number
  /**
   * @type string | undefined
   */
  text?: string
}